Made WAVE_SPAWN_TIME a double instead of an unsigned int.

+++ Time to Defend/src/Settings.h
@@ -17,4 +17,4 @@ constexpr unsigned int MAX_ENEMY_MOVE_TIME = 380;
 static const char* DEFAULT_FONT_PATH = "res/fonts/arial.ttf";
 
 // Time before next wave spawns
-constexpr unsigned int WAVE_SPAWN_TIME = MAX_ENEMY_MOVE_TIME * 2.5;
+constexpr double WAVE_SPAWN_TIME = MAX_ENEMY_MOVE_TIME * 2.5;
